requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
PyMuPDF==1.24.7
Pillow==10.4.0